% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_mt.R
\name{analyze_mt}
\alias{analyze_mt}
\title{analyze_mt}
\usage{
analyze_mt(dt_v_mt, dt_r_mt, dt_criteria, show_progress = TRUE)
}
\arguments{
\item{dt_v_mt}{A data.table of victim profiles (mtDNA)}

\item{dt_r_mt}{A data.table of reference profiles (mtDNA)}

\item{dt_criteria}{A data.table of criteria}

\item{show_progress}{Whether the progress is shown or not}
}
\description{
The function to analyze data for mtDNA
}
