% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_y.R
\name{analyze_y}
\alias{analyze_y}
\title{analyze_y}
\usage{
analyze_y(dt_v_y, dt_r_y, dt_criteria, show_progress = TRUE)
}
\arguments{
\item{dt_v_y}{A data.table of victim profiles (Y-STR)}

\item{dt_r_y}{A data.table of reference profiles (Y-STR)}

\item{dt_criteria}{A data.table of criteria}

\item{show_progress}{Whether the progress is shown or not}
}
\description{
The function to analyze data for Y-STR
}
