% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{change_data_type_af}
\alias{change_data_type_af}
\title{change_data_type_af}
\usage{
change_data_type_af(dt)
}
\arguments{
\item{dt}{The data.table of the allele frequency database}
}
\description{
The function to change data type for the allele frequency database
}
