% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{change_data_type_auto}
\alias{change_data_type_auto}
\title{change_data_type_auto}
\usage{
change_data_type_auto(dt)
}
\arguments{
\item{dt}{The data.table of the autosomal STR database}
}
\description{
The function to change data type for the autosomal STR database
}
