% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{change_data_type_mt}
\alias{change_data_type_mt}
\title{change_data_type_mt}
\usage{
change_data_type_mt(dt)
}
\arguments{
\item{dt}{The data.table of the mtDNA database}
}
\description{
The function to change data type for the mtDNA database
}
