% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{change_data_type_y}
\alias{change_data_type_y}
\title{change_data_type_y}
\usage{
change_data_type_y(dt)
}
\arguments{
\item{dt}{The data.table of the Y-STR database}
}
\description{
The function to change data type for the Y-STR database
}
