% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_error}
\alias{check_error}
\title{check_error}
\usage{
check_error(
  dt_v_auto,
  dt_r_auto,
  dt_af,
  dt_v_y,
  dt_r_y,
  dt_v_mt,
  dt_r_mt,
  dt_rel,
  dt_myu,
  bool_load_v_auto,
  bool_load_r_auto,
  bool_load_af,
  bool_load_v_y,
  bool_load_r_y,
  bool_load_v_mt,
  bool_load_r_mt
)
}
\arguments{
\item{dt_v_auto}{A data.table of victim profiles (autosomal STR)}

\item{dt_r_auto}{A data.table of reference profiles (autosomal STR)}

\item{dt_af}{A data.table of allele frequencies (autosomal STR)}

\item{dt_v_y}{A data.table of victim profiles (Y-STR)}

\item{dt_r_y}{A data.table of reference profiles (Y-STR)}

\item{dt_v_mt}{A data.table of victim profiles (mtDNA)}

\item{dt_r_mt}{A data.table of reference profiles (mtDNA)}

\item{dt_rel}{A data.table of information on relationships}

\item{dt_myu}{A data.table of mutation rates}

\item{bool_load_v_auto}{Whether autosomal STR database for victims has been loaded or not}

\item{bool_load_r_auto}{Whether autosomal STR database for references has been loaded or not}

\item{bool_load_af}{Whether allele frequency database has been loaded or not}

\item{bool_load_v_y}{Whether Y-STR database for victims has been loaded or not}

\item{bool_load_r_y}{Whether Y-STR database for references has been loaded or not}

\item{bool_load_v_mt}{Whether mtDNA database for victims has been loaded or not}

\item{bool_load_r_mt}{Whether mtDNA database for references has been loaded or not}
}
\description{
The function to check loaded data
}
