% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_tree.R
\name{check_tree}
\alias{check_tree}
\title{check_tree}
\usage{
check_tree(persons, sex_all, father_all, mother_all, founder_all)
}
\arguments{
\item{persons}{Names of persons}

\item{sex_all}{Sex of all persons}

\item{father_all}{Fathers of all persons}

\item{mother_all}{Mothers of all persons}

\item{founder_all}{Information on founder/offspring of all persons}
}
\description{
The function to check family tree
}
