% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_auto.R
\name{correct_af_dirichlet}
\alias{correct_af_dirichlet}
\title{correct_af_dirichlet}
\usage{
correct_af_dirichlet(pop_al, pop_freq, unobs_al = numeric(0))
}
\arguments{
\item{pop_al}{A vector of the alleles in the population database}

\item{pop_freq}{A vector of the allele frequencies in the population database}

\item{unobs_al}{A vector of the unobserved alleles in the population database}
}
\description{
The function to correct allele frequencies based on the Dirichlet distribution
}
