% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_combined.R
\name{create_background_color}
\alias{create_background_color}
\title{create_background_color}
\usage{
create_background_color(dt_combined, dt_criteria, dt_rel)
}
\arguments{
\item{dt_combined}{A data.table of the combined data}

\item{dt_criteria}{A data.table of criteria}

\item{dt_rel}{A data.table of information on relationships}
}
\description{
The function to create the sign of the number of candidates (0: Excluded, 1: Inconclusive, 2: Multiple candidates, 3: Identified)
}
