% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_combined.R
\name{create_combined_data}
\alias{create_combined_data}
\title{create_combined_data}
\usage{
create_combined_data(
  dt_result_auto,
  dt_result_y,
  sn_v_y_male,
  sn_r_y_male,
  dt_result_mt,
  dt_criteria,
  dt_rel,
  dt_data_manage
)
}
\arguments{
\item{dt_result_auto}{A data.table of the result for the autosomal STR}

\item{dt_result_y}{A data.table of the result for the Y-STR}

\item{sn_v_y_male}{A character vector of the victim names who are estimated to be male}

\item{sn_r_y_male}{A character vector of the reference names who are estimated to be male}

\item{dt_result_mt}{A data.table of the result for the mtDNA}

\item{dt_criteria}{A data.table of criteria}

\item{dt_rel}{A data.table of information on relationships}

\item{dt_data_manage}{A data.table for data management}
}
\description{
The function to create combined data
}
