% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed_data.R
\name{create_detailed_data_auto}
\alias{create_detailed_data_auto}
\title{create_detailed_data_auto}
\usage{
create_detailed_data_auto(
  dt_v_auto,
  dt_r_auto,
  sn_v_select,
  sn_r_select,
  assumed_rel_select,
  result_selected
)
}
\arguments{
\item{dt_v_auto}{A data.table of victim profiles (autosomal STR)}

\item{dt_r_auto}{A data.table of reference profiles (autosomal STR)}

\item{sn_v_select}{The sample name of the selected victim}

\item{sn_r_select}{The sample name of the selected reference}

\item{assumed_rel_select}{The assumed relationship}

\item{result_selected}{The selected result}
}
\description{
The function to create the detailed data for autosomal STR
}
