% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed_data.R
\name{create_detailed_data_mt}
\alias{create_detailed_data_mt}
\title{create_detailed_data_mt}
\usage{
create_detailed_data_mt(
  dt_v_mt,
  dt_r_mt,
  sn_v_select,
  sn_r_select,
  assumed_rel_select,
  result_selected
)
}
\arguments{
\item{dt_v_mt}{A data.table of victim profiles (mtDNA)}

\item{dt_r_mt}{A data.table of reference profiles (mtDNA)}

\item{sn_v_select}{The sample name of the selected victim}

\item{sn_r_select}{The sample name of the selected reference}

\item{assumed_rel_select}{The assumed relationship}

\item{result_selected}{The selected result}
}
\description{
The function to create the detailed data for mtDNA
}
