% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed_data.R
\name{create_detailed_data_y}
\alias{create_detailed_data_y}
\title{create_detailed_data_y}
\usage{
create_detailed_data_y(
  dt_v_y,
  dt_r_y,
  sn_v_select,
  sn_r_select,
  assumed_rel_select,
  paternal_select,
  result_selected
)
}
\arguments{
\item{dt_v_y}{A data.table of victim profiles (Y-STR)}

\item{dt_r_y}{A data.table of reference profiles (Y-STR)}

\item{sn_v_select}{The sample name of the selected victim}

\item{sn_r_select}{The sample name of the selected reference}

\item{assumed_rel_select}{The assumed relationship}

\item{paternal_select}{The paternal lineage of the selected victim-reference pair}

\item{result_selected}{The selected result}
}
\description{
The function to create the detailed data for Y-STR
}
