% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_combined.R
\name{create_displayed_data}
\alias{create_displayed_data}
\title{create_displayed_data}
\usage{
create_displayed_data(
  dt_combined,
  fltr_type = "default",
  max_data = 10000,
  min_lr = NULL
)
}
\arguments{
\item{dt_combined}{A data.table of the combined data}

\item{fltr_type}{The filtering method}

\item{max_data}{The maximum data displayed}

\item{min_lr}{The minimum LR displayed}
}
\description{
The function to create the displayed data
}
