% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_dt_data_manage}
\alias{create_dt_data_manage}
\title{create_dt_data_manage}
\usage{
create_dt_data_manage(path_pack, init = TRUE)
}
\arguments{
\item{path_pack}{Package path}
}
\description{
The function to create the data.table for setting of data management
}
