% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_dt_other_par}
\alias{create_dt_other_par}
\title{create_dt_other_par}
\usage{
create_dt_other_par(path_pack, init = TRUE)
}
\arguments{
\item{path_pack}{Package path}
}
\description{
The function to create the data.table for other parameters
}
