% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_dt_par_auto}
\alias{create_dt_par_auto}
\title{create_dt_par_auto}
\usage{
create_dt_par_auto(path_pack, init = TRUE)
}
\arguments{
\item{path_pack}{Package path}
}
\description{
The function to create the data.table for parameters of autosomal STR
}
