% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_dt_rel}
\alias{create_dt_rel}
\title{create_dt_rel}
\usage{
create_dt_rel(path_pack)
}
\arguments{
\item{path_pack}{Package path}
}
\description{
The function to create the data.table for information on relationships
}
