% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_combined.R
\name{create_group_cand}
\alias{create_group_cand}
\title{create_group_cand}
\usage{
create_group_cand(dt_combined, background_color)
}
\arguments{
\item{dt_combined}{A data.table of the combined data}

\item{background_color}{The background colors for the displayed table}
}
\description{
The function to create the groups of multiple candidates
}
