% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_combined.R
\name{create_num_cand}
\alias{create_num_cand}
\title{create_num_cand}
\usage{
create_num_cand(dt_combined)
}
\arguments{
\item{dt_combined}{A data.table of the combined data}
}
\description{
The function to create the sign of the number of candidates
}
