% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_auto.R
\name{define_action_myu}
\alias{define_action_myu}
\title{define_action_myu}
\usage{
define_action_myu(dt_rel)
}
\arguments{
\item{dt_rel}{A data.table of information on relationships}
}
\description{
The function to define actions for considering mutation rates
}
