% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detailed_data.R
\name{display_gt}
\alias{display_gt}
\title{display_gt}
\usage{
display_gt(data_gt)
}
\arguments{
\item{data_gt}{Genotype data}
}
\description{
The function to make a vector of genotypes for detailed data
}
