% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_example.R
\name{example_server}
\alias{example_server}
\title{example_server}
\usage{
example_server(id, path_pack)
}
\arguments{
\item{path_pack}{Package path}
}
\description{
The function to create the server module for downloading example files
}
