% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_cand.R
\name{extract_all_cand}
\alias{extract_all_cand}
\title{extract_all_cand}
\usage{
extract_all_cand(dt_display, sn_v_select, sn_r_select)
}
\arguments{
\item{dt_display}{A data.table of the displayed data}

\item{sn_v_select}{The sample name of the selected victim}

\item{sn_r_select}{The sample name of the selected reference}
}
\description{
The function to extract other candidates from dt_display
}
