% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_y.R
\name{judge_male_y}
\alias{judge_male_y}
\title{judge_male_y}
\usage{
judge_male_y(dt_y)
}
\arguments{
\item{dt_y}{A data.table of the Y-STR profiles}
}
\description{
The function to judge whether each individual is male or not based on the Y-STR profiles
}
