% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_tree.R
\name{judge_maternal}
\alias{judge_maternal}
\title{judge_maternal}
\usage{
judge_maternal(p1, p2, id, mid)
}
\arguments{
\item{p1}{The name of person 1}

\item{p2}{The name of person 2}

\item{id}{All names of persons}

\item{mid}{Mothers of all persons}
}
\description{
The function to judge maternal lineage
}
