% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_tree.R
\name{judge_paternal}
\alias{judge_paternal}
\title{judge_paternal}
\usage{
judge_paternal(p1, p2, id, fid, sex)
}
\arguments{
\item{p1}{The name of person 1}

\item{p2}{The name of person 2}

\item{id}{All names of persons}

\item{fid}{Fathers of all persons}

\item{sex}{Sex of all persons}
}
\description{
The function to judge paternal lineage
}
