% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_load_proj.R
\name{load_proj_server}
\alias{load_proj_server}
\title{load_proj_server}
\usage{
load_proj_server(id, session_top, max_data)
}
\description{
The function to create the server module for loading the project
}
