% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_load.R
\name{load_server}
\alias{load_server}
\title{load_server}
\usage{
load_server(id, session_top, rv_criteria, rv_rel, rv_myu, rv_data_manage)
}
\arguments{
\item{session_top}{The top session}

\item{rv_criteria}{The reactive values for criteria}

\item{rv_rel}{The reactive values for information on the relationships}

\item{rv_myu}{The reactive values for mutation rates}

\item{rv_data_manage}{The reactive values for data management}
}
\description{
The function to create the server module for loading files
}
