% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_tree.R
\name{make_deg_display}
\alias{make_deg_display}
\title{make_deg_display}
\usage{
make_deg_display(deg, k1 = -1)
}
\arguments{
\item{deg}{Degree of relationship}

\item{k1}{Pr(IBD = 1)}
}
\description{
The function to make a displayed degree
}
