% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_auto.R
\name{make_dt_af_use}
\alias{make_dt_af_use}
\title{make_dt_af_use}
\usage{
make_dt_af_use(af_list, af_al_list)
}
\arguments{
\item{af_list}{A list of the allele frequencies}

\item{af_al_list}{A list of the alleles}
}
\description{
The function to make a data.table of the allele freuqencies used
}
