% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_auto.R
\name{make_dt_unobs_al}
\alias{make_dt_unobs_al}
\title{make_dt_unobs_al}
\usage{
make_dt_unobs_al(unobs_al_list)
}
\arguments{
\item{unobs_al_list}{A list of the unobserved alleles}
}
\description{
The function to make a data.table of the unobserved alleles
}
