% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_auto.R
\name{make_info_myu}
\alias{make_info_myu}
\title{make_info_myu}
\usage{
make_info_myu(dt_rel)
}
\arguments{
\item{dt_rel}{A data.table of information on relationships}
}
\description{
The function to make information on considering mutations
}
