% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_manual.R
\name{manual_server}
\alias{manual_server}
\title{manual_server}
\usage{
manual_server(id, path_pack)
}
\arguments{
\item{path_pack}{Package path}
}
\description{
The function to create the server module for the manual
}
