% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_general.R
\name{numeric_server}
\alias{numeric_server}
\title{numeric_server}
\usage{
numeric_server(input, output, session, lab, val)
}
\arguments{
\item{lab}{Label}

\item{val}{Value}
}
\description{
The function to create the server module for numericInput
}
