% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_y.R
\name{order_loci_y}
\alias{order_loci_y}
\title{order_loci_y}
\usage{
order_loci_y(dt_v_y, dt_r_y)
}
\arguments{
\item{dt_v_y}{A data.table of victim profiles (Y-STR)}

\item{dt_r_y}{A data.table of reference profiles (Y-STR)}
}
\description{
The function to rearrange in order of loci for Y-STR database
}
