% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{RelSearch-package}
\alias{RelSearch-package}
\title{RelSearch: Open-source software for searching relatives between victim and reference databases}
\description{
RelSearch is open-source software for searching relatives between victim and reference databases. The software is graphical-user-interface written in R language (>= 4.4.0) with GNU General Public License v3.0. It can be applied to autosomal short tandem repeat (STR) markers, Y-STR markers, and mitochondrial DNA sequences commonly used in forensic genetics. For autosomal STR markers, the likelihood ratios (LRs) of each victim-reference pair are calculated considering mutations and drop-out. For Y-STR markers and mitochondrial DNA sequences, the software investigates the number of inconsistencies between victim and reference haplotypes.
}
