% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_result.R
\name{result_server}
\alias{result_server}
\title{result_server}
\usage{
result_server(id, rv_file)
}
\arguments{
\item{rv_file}{The reactive valus for the load data and the analysis results}
}
\description{
The function to create the server module for displaying results
}
