% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_auto.R
\name{set_af}
\alias{set_af}
\title{set_af}
\usage{
set_af(dt_v_auto, dt_r_auto, dt_af, maf)
}
\arguments{
\item{dt_v_auto}{A data.table of victim profiles (autosomal STR)}

\item{dt_r_auto}{A data.table of reference profiles (autosomal STR)}

\item{dt_af}{A data.table of allele frequencies (autosomal STR)}

\item{maf}{The minimum allele frequency}
}
\description{
The function to set allele frequencies
}
