% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_tab_myu.R
\name{tab_myu_server}
\alias{tab_myu_server}
\title{tab_myu_server}
\usage{
tab_myu_server(input, output, session, init_dt_myu, path_pack)
}
\arguments{
\item{init_dt_myu}{The initial data.table of mutation rates}

\item{path_pack}{Package path}
}
\description{
The function to create the server module for mutation rates
}
