% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_par_auto.R
\name{tab_par_auto_server}
\alias{tab_par_auto_server}
\title{tab_par_auto_server}
\usage{
tab_par_auto_server(id, path_pack)
}
\arguments{
\item{path_pack}{Package path}
}
\description{
The function to create the server module for the parameter
}
