% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_rel.R
\name{tab_rel_server}
\alias{tab_rel_server}
\title{tab_rel_server}
\usage{
tab_rel_server(id, path_pack)
}
\arguments{
\item{path_pack}{Package path}
}
\description{
The function to create the server module for information on the relationship
}
