% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_view_dt_af.R
\name{tab_view_dt_af_server}
\alias{tab_view_dt_af_server}
\title{tab_view_dt_af_server}
\usage{
tab_view_dt_af_server(id, rv_file)
}
\description{
The function to create the server module for viewing allele frequency database of the autosomal STR
}
