% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_view_dt_af.R
\name{tab_view_dt_af_ui}
\alias{tab_view_dt_af_ui}
\title{tab_view_dt_af_ui}
\usage{
tab_view_dt_af_ui(id)
}
\description{
The function to create the ui module for viewing allele frequency database of the autosomal STR
}
