% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_view_dt_r_auto.R
\name{tab_view_dt_r_auto_server}
\alias{tab_view_dt_r_auto_server}
\title{tab_view_dt_r_auto_server}
\usage{
tab_view_dt_r_auto_server(id, rv_file)
}
\description{
The function to create the server module for viewing reference database of the autosomal STR
}
