% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_view_dt_r_auto.R
\name{tab_view_dt_r_auto_ui}
\alias{tab_view_dt_r_auto_ui}
\title{tab_view_dt_r_auto_ui}
\usage{
tab_view_dt_r_auto_ui(id)
}
\description{
The function to create the ui module for viewing reference database of the autosomal STR
}
