% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_view_dt_r_y.R
\name{tab_view_dt_r_y_ui}
\alias{tab_view_dt_r_y_ui}
\title{tab_view_dt_r_y_ui}
\usage{
tab_view_dt_r_y_ui(id)
}
\description{
The function to create the ui module for viewing reference database of the Y-STR
}
