% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_view_dt_v_auto.R
\name{tab_view_dt_v_auto_server}
\alias{tab_view_dt_v_auto_server}
\title{tab_view_dt_v_auto_server}
\usage{
tab_view_dt_v_auto_server(id, rv_file)
}
\arguments{
\item{rv_file}{The reactive valus for the load data and the analysis results}
}
\description{
The function to create the server module for viewing victim database of the autosomal STR
}
