% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_view_dt_v_mt.R
\name{tab_view_dt_v_mt_ui}
\alias{tab_view_dt_v_mt_ui}
\title{tab_view_dt_v_mt_ui}
\usage{
tab_view_dt_v_mt_ui(id)
}
\description{
The function to create the ui module for viewing victim database of the mtDNA
}
