% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_view_dt_v_y.R
\name{tab_view_dt_v_y_server}
\alias{tab_view_dt_v_y_server}
\title{tab_view_dt_v_y_server}
\usage{
tab_view_dt_v_y_server(id, rv_file)
}
\description{
The function to create the server module for viewing victim database of the Y-STR
}
