% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_combined.R
\name{create_background_color}
\alias{create_background_color}
\title{create_background_color}
\usage{
create_background_color(dt_combined, index_warning)
}
\arguments{
\item{dt_combined}{A data.table of the combined data}

\item{index_warning}{Indices of not-supporting paternal or maternal lineage with LR >= 100}
}
\description{
The function to create the sign of the number of candidates
}
