% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_combined.R
\name{create_background_color}
\alias{create_background_color}
\title{create_background_color}
\usage{
create_background_color(dt_combined, index_change_yellow, index_change_red)
}
\arguments{
\item{dt_combined}{A data.table of the combined data}

\item{index_change_yellow}{Indices of inconclusive paternal or maternal lineages with supportive LR}

\item{index_change_red}{Indices of excluded paternal or maternal lineages with supportive LR}
}
\description{
The function to create the sign of the number of candidates (0: Excluded, 1: Inconclusive, 2: Multiple candidates, 3: Identified)
}
