% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_combined.R
\name{create_combined_data}
\alias{create_combined_data}
\title{create_combined_data}
\usage{
create_combined_data(dt_result_auto, dt_result_y, dt_result_mt, dt_rel)
}
\arguments{
\item{dt_result_auto}{A data.table of the result for the autosomal STR}

\item{dt_result_y}{A data.table of the result for the Y-STR}

\item{dt_result_mt}{A data.table of the result for the mtDNA}

\item{dt_rel}{A data.table of information on relationships}
}
\description{
The function to create combined data
}
