% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{create_dt_myu}
\alias{create_dt_myu}
\title{create_dt_myu}
\usage{
create_dt_myu(path_pack)
}
\arguments{
\item{path_pack}{Package path}
}
\description{
The function to create the data.table for mutation rates
}
