% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_load.R
\name{load_server}
\alias{load_server}
\title{load_server}
\usage{
load_server(
  id,
  session_top,
  rv_criteria,
  rv_rel,
  rv_myu,
  rv_par_auto,
  keep_min_lr,
  max_data
)
}
\description{
The function to create the server module for loading files
}
