% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{relsearch-package}
\alias{relsearch}
\alias{relsearch-package}
\title{relsearch: Open-source software for searching relatives between victim and reference databases}
\description{
relsearch is open-source software for screening relatives between query and reference databases. The software is graphical-user-interface written in R language (>= 4.2.0) with GNU General Public License v3.0. It can be applied to autosomal short tandem repeat (STR) markers, Y-STR markers, and mitochondrial DNA sequences commonly used in forensic genetics. For autosomal STR markers, likelihood ratios (LRs) of each query-reference pair are calculated considering mutation and drop-out. For Y-STR markers and mitochondrial DNA sequences, the software investigates the number of inconsistencies between query and reference haplotypes.
}
