% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_criteria.R
\name{tab_criteria_server}
\alias{tab_criteria_server}
\title{tab_criteria_server}
\usage{
tab_criteria_server(id, path_pack)
}
\arguments{
\item{path_pack}{Package path}
}
\description{
The function to create the server module for criteria
}
