% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_data_management.R
\name{tab_data_manage_server}
\alias{tab_data_manage_server}
\title{tab_data_manage_server}
\usage{
tab_data_manage_server(id, path_pack)
}
\arguments{
\item{path_pack}{Package path}
}
\description{
The function to create the server module for data management
}
